Use a sensible entropy source for generateGenesis{Data,Config}
It doesn't really matter that much since these are only ever used for
generating test genesis files and test keys, however it's still better
to do the right thing even if not strictly necessary.

So to use a proper entropy source we need to use MonadRandom and then
instantiate it with a proper source. So we generalise
generateGenesis{Data,Config} over MonadRandom, renaming them to
generateGenesis{Data,Config}WithEntropy. Then we specialise the original
ones to IO and us the Cardano.Crypto.Random.runSecureRandom as the
entropy source which is suitable for the purpose of key generation.

This means we can still use a pure deterministic "entropy" source for
generating mock/dummy examples, but can also use a proper entropy source
that's suitable for key generation (again, even though that's not
strictly necessary since this whole module is only for test
configurations).