Switch secure entropy source from openssl to system entropy
We only use this for key generation so it is simpler and probably better
to get entropy straight from the operating system, rather than use
OpenSSL's entropy pool. The OpenSSL one is better suited to generating
lots of keys quickly, but we do not need that.

This also lets us drop the dependency on OpenSSL which is a portability
benefit.